% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yin-DFT-Distances-R.r
\name{testOfPSDiff}
\alias{testOfPSDiff}
\title{A Statistical Test based on the varianceGamma Distribution for the 
  difference in Power spectra}
\usage{
testOfPSDiff(PS1, PS2)
}
\description{
A Statistical Test based on the varianceGamma Distribution for the 
  difference in Power spectra
}
