% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yin-DFT-Distances-R.r
\name{testOfPSRatio}
\alias{testOfPSRatio}
\title{A Statistical Test based on the F Distribution for the ratio of two Power 
  Spectra}
\usage{
testOfPSRatio(PS1, PS2)
}
\description{
A Statistical Test based on the F Distribution for the ratio of two Power 
  Spectra
}
